crud

+++ config/data/oil/models.py
@@ -16,7 +16,7 @@ class OilPurchase(TimeStampModel):
     price_usd = models.DecimalField(decimal_places=2, max_digits=10)
     amount_uzs = models.DecimalField(decimal_places=2, max_digits=10)
     amount_usd = models.DecimalField(decimal_places=2, max_digits=10)
-    oil_volume = models.PositiveIntegerField(help_text="Oil volume in liters")
+    oil_volume = models.FloatField(help_text="Oil volume in liters")
 
     def __str__(self):
         return f"{self.oil.oil_name} - {self.oil_volume} L"
